+++ crates/zksync/core/src/vm/mod.rs
@@ -4,6 +4,8 @@ mod farcall;
 mod runner;
 mod storage_view;
 mod tracer;
+mod tx;
 
 pub use runner::{balance, call, code_hash, create, encode_create_params, nonce, transact};
 pub use tracer::CheatcodeTracerContext;
+pub use tx::split_tx_by_factory_deps;
